+++ PythonAPI/carla/src/Client.cpp
@@ -11,6 +11,7 @@
 #include <boost/python/stl_iterator.hpp>
 
 namespace ctm = carla::traffic_manager;
+namespace cg = carla::geom;
 
 static void SetTimeout(carla::client::Client &client, double seconds) {
   client.SetTimeout(TimeDurationFromSeconds(seconds));
@@ -185,6 +186,29 @@ void export_client() {
     .def_readwrite("enable_pedestrian_navigation", &rpc::OpendriveGenerationParameters::enable_pedestrian_navigation)
   ;
 
+#define REPLAY_FILE_WITHOUT_GIL(fn) +[]( \
+        cc::Client &self, \
+        std::string name, \
+        double time_start, \
+        double duration, \
+        uint32_t follow_id, \
+        bool replay_sensors, \
+        bool replay_weather, \
+        cg::Transform offset, \
+        std::string map_override) { \
+      carla::PythonUtil::ReleaseGIL unlock; \
+      return self.fn(name, time_start, duration, follow_id, replay_sensors, replay_weather, offset, map_override); \
+    }, \
+    ( \
+      arg("name"), \
+      arg("time_start"), \
+      arg("duration"), \
+      arg("follow_id"), \
+      arg("replay_sensors")=false, \
+      arg("replay_weather")=false, \
+      arg("offset")=cg::Transform(cg::Location(cg::Vector3D(-10, 0, 5)), cg::Rotation(-25, 0, 0)), \
+      arg("map_override")="")
+
   class_<cc::Client>("Client",
       init<std::string, uint16_t, size_t>((arg("host")="127.0.0.1", arg("port")=2000, arg("worker_threads")=0u)))
     .def("set_timeout", &::SetTimeout, (arg("seconds")))
@@ -201,12 +225,12 @@ void export_client() {
     .def("generate_opendrive_world", CONST_CALL_WITHOUT_GIL_3(cc::Client, GenerateOpenDriveWorld, std::string,
         rpc::OpendriveGenerationParameters, bool), (arg("opendrive"), arg("parameters")=rpc::OpendriveGenerationParameters(),
         arg("reset_settings")=true))
-    .def("start_recorder", CALL_WITHOUT_GIL_2(cc::Client, StartRecorder, std::string, bool), (arg("name"), arg("additional_data")=false))
+    .def("start_recorder", CALL_WITHOUT_GIL_3(cc::Client, StartRecorder, std::string, bool, bool), (arg("name"), arg("additional_data")=false, arg("stop_replayer")=true))
     .def("stop_recorder", &cc::Client::StopRecorder)
     .def("show_recorder_file_info", CALL_WITHOUT_GIL_2(cc::Client, ShowRecorderFileInfo, std::string, bool), (arg("name"), arg("show_all")))
     .def("show_recorder_collisions", CALL_WITHOUT_GIL_3(cc::Client, ShowRecorderCollisions, std::string, char, char), (arg("name"), arg("type1"), arg("type2")))
     .def("show_recorder_actors_blocked", CALL_WITHOUT_GIL_3(cc::Client, ShowRecorderActorsBlocked, std::string, double, double), (arg("name"), arg("min_time"), arg("min_distance")))
-    .def("replay_file", CALL_WITHOUT_GIL_5(cc::Client, ReplayFile, std::string, double, double, uint32_t, bool), (arg("name"), arg("time_start"), arg("duration"), arg("follow_id"), arg("replay_sensors")=false))
+    .def("replay_file", REPLAY_FILE_WITHOUT_GIL(ReplayFile))
     .def("stop_replayer", &cc::Client::StopReplayer, (arg("keep_actors")))
     .def("set_replayer_time_factor", &cc::Client::SetReplayerTimeFactor, (arg("time_factor")))
     .def("set_replayer_ignore_hero", &cc::Client::SetReplayerIgnoreHero, (arg("ignore_hero")))